MAINT: Split into decorators and exceptions module
